# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "Config" : {
        "properties" : {
          "goalLimit" : {
            "format" : "int32",
            "type" : "integer"
          },
          "matchMinutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "plannedRounds" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "pointsDraw" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pointsLoss" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pointsWin" : {
            "format" : "int32",
            "type" : "integer"
          },
          "randomRounds" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tableCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "Event" : {
        "properties" : {
          "id" : {
            "description" : "Kennung des Events. Beim Neuverbinden als Last-Event-ID senden, dann werden verpasste Events nachgeliefert",
            "type" : "string"
          },
          "matchId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "type" : {
            "description" : "TOURNAMENT_CHANGED, PARTICIPANTS_CHANGED, ROUND_STARTED, MATCHES_CHANGED, RANKING_CHANGED oder RESYNC (Stand komplett neu laden)",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Match" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "mySide" : {
            "description" : "Auf welcher Seite der Aufrufer steht (auch als Einspringer), null wenn gar nicht",
            "enum" : [ "A", "B" ],
            "type" : [ "string", "null" ]
          },
          "permissions" : {
            "$ref" : "#/components/schemas/MatchPermissions",
            "description" : "Was der Aufrufer mit diesem Match jetzt tun darf"
          },
          "position" : {
            "description" : "Position in der Warteschlange der Runde, beginnend bei 1",
            "format" : "int32",
            "type" : "integer"
          },
          "result" : {
            "$ref" : "#/components/schemas/Result",
            "description" : "Eingetragenes oder festgelegtes Ergebnis, null solange keines vorliegt",
            "type" : "null"
          },
          "round" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "QUEUED", "ON_TABLE", "RESULT_ENTERED", "DISPUTED", "CONFIRMED" ],
            "type" : "string"
          },
          "table" : {
            "description" : "Tischnummer, null solange das Match wartet",
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "teamA" : {
            "$ref" : "#/components/schemas/Team"
          },
          "teamB" : {
            "$ref" : "#/components/schemas/Team"
          }
        },
        "type" : "object"
      },
...
      "Participant" : {
        "properties" : {
          "player" : {
            "$ref" : "#/components/schemas/Player"
          },
          "registeredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "PAUSED", "WITHDRAWN" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PlanTournament" : {
        "properties" : {
          "config" : {
            "$ref" : "#/components/schemas/Config",
            "description" : "Optional, ohne Angabe gelten die Standardwerte"
          },
          "date" : {
            "example" : "2026-10-05",
            "format" : "date",
            "type" : "string"
          },
          "name" : {
            "example" : "Sommerturnier",
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "date", "name" ],
        "type" : "object"
      },
      "Player" : {
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Round" : {
        "properties" : {
          "complete" : {
            "description" : "Alle Ergebnisse der Runde sind bestätigt",
            "type" : "boolean"
          },
          "matches" : {
            "description" : "In der Reihenfolge der Auslosung, also der Warteschlange für die Tische",
            "items" : {
              "$ref" : "#/components/schemas/Match"
            },
            "type" : "array"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "Slot" : {
        "properties" : {
          "player" : {
            "$ref" : "#/components/schemas/Player",
            "type" : "null"
          },
          "standIn" : {
            "$ref" : "#/components/schemas/Player",
            "description" : "Nur bei DUMMY: wer tatsächlich spielt, ohne Punkte zu bekommen",
            "type" : "null"
          },
          "type" : {
            "enum" : [ "PLAYER", "DUMMY" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Team" : {
        "properties" : {
          "members" : {
            "items" : {
              "$ref" : "#/components/schemas/Slot"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "Tournament" : {
        "properties" : {
          "config" : {
            "$ref" : "#/components/schemas/Config"
          },
          "currentRound" : {
            "$ref" : "#/components/schemas/Round",
            "description" : "Die aktuelle Runde, null solange noch keine gestartet wurde",
            "type" : "null"
          },
          "date" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "participants" : {
            "items" : {
              "$ref" : "#/components/schemas/Participant"
            },
            "type" : "array"
          },
          "roundCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PLANNED", "RUNNING", "FINISHED" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "TournamentSummary" : {
        "properties" : {
          "date" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "participantCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "roundCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PLANNED", "RUNNING", "FINISHED" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access-Token des OIDC-Providers einfügen, ohne das Präfix `Bearer`.",
        "scheme" : "bearer",
        "type" : "http"
      },
      "oidc" : {
        "description" : "Anmeldung über den OIDC-Provider (Authorization Code mit PKCE).",
        "openIdConnectUrl" : "http://localhost:8180/realms/kickertool/.well-known/openid-configuration",
        "type" : "openIdConnect"
      }
    }
  },
  "info" : {
    "description" : "REST-API für Kicker-Turniere (2 gegen 2, wechselnde Partner).\n\n**Anmeldung:** Registrierung und Login übernimmt der OIDC-Provider. Die API prüft nur das Access-Token. Über **Authorize** kann man sich direkt anmelden (`oidc`) oder ein Token einfügen (`bearerAuth`). Admin-Endpunkte verlangen die konfigurierte Admin-Rolle im Token, ein normaler Spieler bekommt dort 403.\n\n**Fehler** kommen als `application/problem+json`.\n\n**Live-Updates:** `GET /api/tournaments/{id}/events` liefert Server-Sent Events. Swagger UI kann diesen Stream nicht darstellen, dafür `curl -N` mit Bearer-Header nutzen.",
    "title" : "Kickertool API",
    "version" : "0.1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/tournaments" : {
      "get" : {
        "description" : "Neueste zuerst.",
        "operationId" : "list",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TournamentSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Ungültige Eingabe, z. B. Pflichtfeld fehlt oder Ergebnis über dem Tor-Limit"
          },
          "401" : {
            "description" : "Kein oder ungültiges Token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Keine Berechtigung, z. B. kein Admin oder nicht Teil des Matches"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Nicht gefunden"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Passt nicht zum aktuellen Stand, z. B. falscher Turnierstatus oder Runde nicht abgeschlossen"
          }
        },
        "summary" : "Turniere auflisten",
        "tags" : [ "Turniere" ]
      },
      "post" : {
        "description" : "**Nur Admin.** Ohne `config` gelten die Standardwerte (1 Tisch, 3/1/0 Punkte, 10 Tore, 5 Minuten, 2 Zufallsrunden). Danach können sich Spieler anmelden.",
        "operationId" : "plan",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PlanTournament"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Tournament"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Ungültige Eingabe, z. B. Pflichtfeld fehlt oder Ergebnis über dem Tor-Limit"
          },
          "401" : {
            "description" : "Kein oder ungültiges Token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Keine Berechtigung, z. B. kein Admin oder nicht Teil des Matches"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Nicht gefunden"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Passt nicht zum aktuellen Stand, z. B. falscher Turnierstatus oder Runde nicht abgeschlossen"
          }
        },
        "summary" : "Turnier planen",
        "tags" : [ "Turniere" ]
      }
    },
...
    "/api/tournaments/{tournamentId}/events" : {
      "get" : {
        "description" : "Server-Sent Events. Sendet `connected` und danach bei Änderungen die Events `TOURNAMENT_CHANGED`, `PARTICIPANTS_CHANGED`, `ROUND_STARTED`, `MATCHES_CHANGED` und `RANKING_CHANGED` (JSON mit `id`, `type`, `matchId`, `occurredAt`, ohne Nutzdaten). Clients laden danach den neuen Stand über die normalen Endpunkte.\n\n**Anmeldung:** mit dem Bearer-Token im `Authorization`-Header oder mit einem Ticket in `ticket`.\n\n**Nachliefern:** Jedes Event hat eine `id`. Wer sich mit der `id` des letzten empfangenen Events in `Last-Event-ID` (oder, wo kein Header möglich ist, in `lastEventId`) neu verbindet, der Browser macht das bei einer Wiederverbindung desselben `EventSource` von allein, bekommt die verpassten Events nachgeliefert. Geht das nicht mehr (zu lange weg, Server neu gestartet), kommt ein einzelnes `RESYNC`: Dann den Stand komplett neu laden.\n\nIn Swagger UI nicht darstellbar, lokal z. B. mit `curl -N -H \"Authorization: Bearer ...\"`.",
        "operationId" : "stream",
        "parameters" : [ {
          "in" : "path",
          "name" : "tournamentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Kennung des letzten empfangenen Events, für das Nachliefern verpasster Events",
          "in" : "header",
          "name" : "Last-Event-ID",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Ticket statt Bearer-Token, siehe `POST …/events/ticket`",
          "in" : "query",
          "name" : "ticket",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Wie `Last-Event-ID`, für Clients, die keinen Header setzen können: ein neues `EventSource`-Objekt kann das nicht. Der Header hat Vorrang.",
          "in" : "query",
          "name" : "lastEventId",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/event-stream" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Event"
                }
              }
            },
            "description" : "Dauerhafte Verbindung mit Events"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Ungültige Eingabe, z. B. Pflichtfeld fehlt oder Ergebnis über dem Tor-Limit"
          },
          "401" : {
            "description" : "Kein oder ungültiges Token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Keine Berechtigung, z. B. kein Admin oder nicht Teil des Matches"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Nicht gefunden"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "properties" : {
                    "detail" : {
                      "example" : "Runde 1 ist noch nicht abgeschlossen",
                      "type" : "string"
                    },
                    "instance" : {
                      "type" : "string"
                    },
                    "status" : {
                      "example" : 409,
                      "format" : "int32",
                      "type" : "integer"
                    },
                    "title" : {
                      "example" : "Conflict",
                      "type" : "string"
                    },
                    "type" : {
                      "example" : "about:blank",
                      "type" : "string"
                    }
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "Passt nicht zum aktuellen Stand, z. B. falscher Turnierstatus oder Runde nicht abgeschlossen"
          }
        },
        "summary" : "Live-Updates abonnieren",
        "tags" : [ "Live-Updates" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  }, {
    "oidc" : [ ]
  } ],
...
}